# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MeetSpace Video Meeting API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface MeetingParticipant {
  userId: string;
  name: string;
  joinedAt: string;
  /** @nullable */
  leftAt?: string | null;
}

export type MeetingStatus = typeof MeetingStatus[keyof typeof MeetingStatus];


export const MeetingStatus = {
  scheduled: 'scheduled',
  active: 'active',
  ended: 'ended',
} as const;

export interface Meeting {
  id: string;
  title: string;
  hostId: string;
  /** @nullable */
  hostName?: string | null;
  roomCode: string;
  status: MeetingStatus;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  duration?: number | null;
  participants?: MeetingParticipant[];
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  recordingUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface MeetingInput {
  /** @minLength 1 */
  title: string;
  scheduledAt?: string;
}

export interface MeetingPatch {
  /** @minLength 1 */
  title?: string;
  scheduledAt?: string;
}

export interface MeetingSummary {
  meetingId: string;
  /** @nullable */
  summary: string | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  duration?: number | null;
  participantCount?: number;
}

export interface MeetingStats {
  totalMeetings: number;
  totalDuration: number;
  upcomingCount: number;
  hostedCount: number;
}

export interface Recording {
  id: string;
  meetingId: string;
  meetingTitle: string;
  url: string;
  /** @nullable */
  publicId?: string | null;
  /** @nullable */
  duration?: number | null;
  createdAt: string;
}