# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MeetSpace Video Meeting API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: meetings
    description: Meeting operations
  - name: users
    description: User operations
  - name: recordings
    description: Recording operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List user meetings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"

    post:
      operationId: createMeeting
      tags: [meetings]
      summary: Create a meeting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings/{meetingId}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get meeting by ID
      security:
        - bearerAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateMeeting
      tags: [meetings]
      summary: Update a meeting
      security:
        - bearerAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"

    delete:
      operationId: deleteMeeting
      tags: [meetings]
      summary: Delete a meeting
      security:
        - bearerAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content

  /meetings/{meetingId}/end:
    post:
      operationId: endMeeting
      tags: [meetings]
      summary: End a meeting
      security:
        - bearerAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"

  /meetings/{meetingId}/summary:
    get:
      operationId: getMeetingSummary
      tags: [meetings]
      summary: Get AI summary for a meeting
      security:
        - bearerAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingSummary"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings/stats:
    get:
      operationId: getMeetingStats
      tags: [meetings]
      summary: Get meeting stats for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingStats"

  /recordings:
    get:
      operationId: listRecordings
      tags: [recordings]
      summary: List recordings for user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"

  /recordings/{recordingId}:
    get:
      operationId: getRecording
      tags: [recordings]
      summary: Get a recording by ID
      security:
        - bearerAuth: []
      parameters:
        - name: recordingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteRecording
      tags: [recordings]
      summary: Delete a recording
      security:
        - bearerAuth: []
      parameters:
        - name: recordingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    User:
      type: object
      required: [id, name, email, avatarUrl, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    MeetingParticipant:
      type: object
      required: [userId, name, joinedAt]
      properties:
        userId:
          type: string
        name:
          type: string
        joinedAt:
          type: string
        leftAt:
          type: ["string", "null"]

    Meeting:
      type: object
      required: [id, title, hostId, roomCode, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        hostId:
          type: string
        hostName:
          type: ["string", "null"]
        roomCode:
          type: string
        status:
          type: string
          enum: [scheduled, active, ended]
        scheduledAt:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        endedAt:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/MeetingParticipant"
        transcript:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        recordingUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    MeetingInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        scheduledAt:
          type: string

    MeetingPatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        scheduledAt:
          type: string

    MeetingSummary:
      type: object
      required: [meetingId, summary]
      properties:
        meetingId:
          type: string
        summary:
          type: ["string", "null"]
        transcript:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        participantCount:
          type: integer

    MeetingStats:
      type: object
      required: [totalMeetings, totalDuration, upcomingCount, hostedCount]
      properties:
        totalMeetings:
          type: integer
        totalDuration:
          type: integer
        upcomingCount:
          type: integer
        hostedCount:
          type: integer

    Recording:
      type: object
      required: [id, meetingId, meetingTitle, url, createdAt]
      properties:
        id:
          type: string
        meetingId:
          type: string
        meetingTitle:
          type: string
        url:
          type: string
        publicId:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        createdAt:
          type: string