# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MeetSpace Video Meeting API
 * OpenAPI spec version: 0.1.0
 */
import type { MeetingParticipant } from './meetingParticipant';
import type { MeetingStatus } from './meetingStatus';

export interface Meeting {
  id: string;
  title: string;
  hostId: string;
  /** @nullable */
  hostName?: string | null;
  roomCode: string;
  status: MeetingStatus;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  duration?: number | null;
  participants?: MeetingParticipant[];
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  recordingUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}