# lib/api-zod/src/generated/types/meetingInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MeetSpace Video Meeting API
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingInput {
  /** @minLength 1 */
  title: string;
  scheduledAt?: string;
}