# lib/api-zod/src/generated/types/meetingStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MeetSpace Video Meeting API
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingStats {
  totalMeetings: number;
  totalDuration: number;
  upcomingCount: number;
  hostedCount: number;
}